# Improved performance of value lookup in AvaloniaObject's ValueStore