```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17763
Intel Core i7-4771 CPU 3.50GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.502
  [Host]     : .NET Core 2.0.9 (CoreCLR 4.6.26614.01, CoreFX 4.6.26614.01), 64bit RyuJIT  [AttachedDebugger]
  DefaultJob : .NET Core 2.0.9 (CoreCLR 4.6.26614.01, CoreFX 4.6.26614.01), 64bit RyuJIT
```


Before:

Method|Mean|Error|StdDev|Gen 0|Gen 1|Allocated
---|---|---|---|---|---|---
ClearAndSetIntProperty|530 ns|3.795 ns |3.364 ns|0.1392||584 B
BindIntProperty|36,480 ns|272 ns|255 ns|13.3667||56069 B
Remeasure_Half|3.452 ms|0.0379 ms|0.0336 ms|312.5|62.5|1.39 MB


After:
        
Method|Mean|Error|StdDev|Gen 0|Gen 1|Allocated
---|---|---|---|---|---|---
ClearAndSetIntProperty|467 ns|4.335 ns|4.055 ns|0.1392||584 B
BindIntProperty|35,874 ns|364 ns|322 ns|13.3667||56069 B
Remeasure_Half|3.203 ms|0.0267 ms|0.0250 ms|312.5| 62.5|1.39 MB