@mstr2 2019-01-04
Another benchmark: This test consists of clicking on the Calendar page in ControlCatalog, and then repeating a 15-second sequence of interactions (zooming out to months view on all six calendars, then zooming in on the next month on all six calendars). The exact sequence of method calls is not stable between successive test runs, but the execution time seems to be consistently lower in the second case.

Before:
![image](https://user-images.githubusercontent.com/43553916/50669434-7fe1f700-0fc5-11e9-91c8-8d83f23f34d9.png)

After:
![image](https://user-images.githubusercontent.com/43553916/50669443-8b352280-0fc5-11e9-9cb2-c87da1b94b27.png)

Calls to SetValue did not account for any significant CPU time (~1 ms in both cases).