@mstr2 2019-01-09
> Out of curiosity, how much performance does the manually unrolled loop actually add?

I've run 12 ClearAndSetIntProperty benchmarks, and the mean difference is 469 ns (unrolled) and 496 ns (for-loop).