@ahopper 2019-01-10
would it be faster for long lists if the binary search fell back to the linear search for the last few terms?