@mstr2 2019-01-16
> would it be faster for long lists if the binary search fell back to the linear search for the last few terms?

The binary search effectively is a linear search for the last three items. I'm not sure if it's worthwhile to make it more compliated than that.