@ahopper 2019-01-16
> > would it be faster for long lists if the binary search fell back to the linear search for the last few terms?
> 
> The binary search effectively is a linear search for the last three items. I'm not sure if it's worthwhile to make it more compliated than that.

my gut reaction is that if 16 items is the optimum cutoff at the first stage it would also roughly be the same in the binary search, it may well then change the balance to a smaller number if the binary option gets faster for mid range values, just guessing and happy to be proved wrong by measurement:)