@mstr2 2019-01-16
@ahopper True, but I think this level of fine-tuning would necessitate a set of more meaningful tests. The length of the linear first stage should actually depend on how far, on average, the linear search progresses until it finds the value. This largely depends on the types of properties that are stored in the ValueStore. Performance also depends on specifics of CPU cache hits/misses. So I agree with your idea, but suggest to make this a future optimization.