@ahopper 2019-01-16
@mstr2 I did some tests a while ago and as far as I can remember the average size of a valuestore on the calendar page of control catalog was 5 ish which probably explains why your solution is better in profiling than in the simple benchmark as it detects misses faster than Dictionary.  As we have control of the property ids we could in future pick the most common to appear at lower numbers to optimise this approach.