@grokys 2019-01-26
Thanks @mstr2 - every little performance improvement in this area helps!